7 failures and 4 errors for spec harness

+++ db/migrate/20160105003838_create_invoice_items.rb
@@ -2,7 +2,7 @@ class CreateInvoiceItems < ActiveRecord::Migration
   def change
     create_table :invoice_items do |t|
       t.integer  :quantity
-      t.integer  :unit_price
+      t.decimal  :unit_price
 
       t.timestamps null: false
     end
